.history/requirements_20200729094731.txt
blinker==1.4
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
cryptography==2.9.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.1
